requirements.txt
requests==2.23.0
beautifulsoup4==4.9.0
python_dateutil==2.8.1
gspread~=3.6.0
python-dateutil~=2.8.1
oauth2client~=4.1.3